requirements.txt
accelerate==0.21.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.41.0
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
blis==0.7.11
braceexpand==0.1.7
cachetools==5.5.0
catalogue==1.0.2
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
cmake==3.30.4
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
decorator==4.4.2
decord==0.6.0
diffusers==0.25.0
distlib==0.3.9
einops==0.8.0
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.25.2
identify==2.6.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.8
marisa-trie==1.2.1
markdown-it-py==2.2.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
narwhals==1.9.3
networkx==3.4.1
nltk==3.9.1
nodeenv==1.9.1
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.10.0.84
opendatasets==0.1.22
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
plac==1.1.3
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
pre_commit==4.0.1
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.12
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
stack-data==0.6.3
starlette==0.39.2
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
text-unidecode==1.3
thinc==7.4.6
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.9
tokenizers==0.14.1
toml==0.10.2
tomli==2.2.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.34.0
triton==2.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.31.1
virtualenv==20.26.6
wasabi==0.10.1
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
websockets==13.1
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
